pyproject.toml
[tool.poetry]
name = "news-scraper"
version = "0.1.0"
description = ""
authors = ["binsarjr <binsarjr121@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scrapy = "^2.11.2"
scrapy-useragents = "^0.0.1"
scrapy-playwright = "^0.0.44"
playwright = "^1.55.0"
dateparser = "^1.2.0"
bs4 = "^0.0.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pandas = "^2.2.3"
numpy = "^2.1.3"
joblib = "^1.4.2"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
spacy = "^3.7.6"
geopy = "^2.4.1"
tqdm = "^4.67.1"
geopandas = "^1.1.1"
scikit-learn = "^1.7.2"
google-generativeai = "^0.8.5"
shapely = "^2.0.6"
openai = "^1.57.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
